requirements.txt
flask==3.0.2
onnxruntime==1.19.2     # or just:  onnxruntime
pillow==10.3.0
numpy
